requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.5.1
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
discord.py==2.3.2
edge-tts==6.1.5
elevenlabs==0.2.21
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fairseq==0.12.2
faiss-cpu==1.7.2
fastapi==0.88.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.10.0
flatbuffers==23.5.9
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
google-auth==2.16.2
google-auth-oauthlib==1.0.0
gradio==3.34.0
gradio_client==0.2.10
grpcio==1.51.3
gspread==5.10.0
gTTS==2.3.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.0.7
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.0.0
ipython==8.15.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonschema==4.17.3
kiwisolver==1.4.4
librosa==0.9.1
linkify-it-py==1.0.3
llvmlite==0.39.1
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.1
mega.py==1.0.8
mpmath==1.2.1
multidict==6.0.3
networkx==2.8.8
noisereduce==2.0.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.0.6
onnx==1.14.0
onnxconverter-common==1.13.0
onnxruntime-gpu==1.14.1
orjson==3.8.3
packaging==22.0
pandas==1.5.2
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.3.0
pooch==1.6.0
portalocker==2.6.0
praat-parselmouth==0.4.2
praatio==6.0.1
prompt-toolkit==3.0.39
protobuf==4.22.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.2
PySimpleGUI==4.60.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.6
pywin32==306
pyworld==0.3.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.2.0
scipy==1.9.3
semantic-version==2.10.0
six==1.16.0
skl2onnx==1.14.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.11.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.22.0
stftpitchshift==1.5.1
sympy==1.11.1
tabulate==0.9.0
tenacity==5.1.5
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.1
threadpoolctl==3.1.0
toolz==0.12.0
torch==2.2.0.dev20230915+cu121
torchaudio==2.2.0.dev20230915+cu121
torchcrepe==0.0.19
torchgen==0.0.1
torchvision==0.17.0.dev20230915+cu121
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
wget==3.2
yarl==1.8.2
zipp==3.15.0